Refactor calendar listings
Creates a new class for calendar lists, now forcing the creating of all
chosen fields, whether or not they are returned by the API. Accordingly,
this refactors the actual list functions to play nice with the class.

Pushes an additional infix utility for handling unwanted (predefinable
as something else) NA values in a vector. Let's see if it works out in
the long run, or if this just ignores an existing (and better) fix...