Address failing builds due to strict tibble refs
Travis builds seemed to be failing because (at least, I suspect) tibbles
were being passed to as.body(). Strangely, nothing fails interactively,
but it looks like when we'd call df$xxx where xxx does not exist as a
column, we get an error rather than the expected NULL (the error is to
be expected with the strictness of tibbles).

I'm unsure what exactly causes the tibble issue, so this removes the
readr suggest (only being used in testing), replaces it with a regular
read.csv(), and changes the as.body dispatch to a regular data.frame.

Not entirely sure this will work, but more annoyingly, I can't reproduce
the errors outside of a Travis build...